fix(client): restore smooth activity indicators
