+++ client/app/lib/features/session_detail/widgets/background_tasks_header.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:theme_prego/module_prego.dart";
 import "../../../core/extensions/build_context_x.dart";
-import "../../../core/widgets/throttled_activity_indicator.dart";
+import "../../../core/widgets/isolated_activity_indicator.dart";
 
 /// Tappable header row of the background-tasks card. Shows a spinner +
 /// "N Tasks Running" (or a check + "Completed") and a chevron that rotates as
@@ -38,7 +38,7 @@ class BackgroundTasksHeader extends StatelessWidget {
               heightFactor: 1,
               child: SizedBox.square(
                 dimension: 16,
-                child: ThrottledActivityIndicator(
+                child: IsolatedActivityIndicator(
                   strokeWidth: 2,
                   color: prego.colors.bgBrandSolid,
                 ),
