+++ client/app/lib/features/session_detail/widgets/prompt_input.dart
@@ -13,7 +13,7 @@ import "../../../core/constants.dart";
 import "../../../core/di/injection.dart";
 import "../../../core/extensions/build_context_x.dart";
 import "../../../core/widgets/command_picker_sheet.dart";
-import "../../../core/widgets/throttled_activity_indicator.dart";
+import "../../../core/widgets/isolated_activity_indicator.dart";
 
 enum _VoiceState { idle, recording, transcribing }
 
@@ -536,7 +536,7 @@ class _TranscribingIndicator extends StatelessWidget {
           SizedBox(
             width: 16,
             height: 16,
-            child: ThrottledActivityIndicator(
+            child: IsolatedActivityIndicator(
               strokeWidth: 2,
               color: prego.colors.bgBrandSolid,
             ),
