+++ client/app/lib/features/session_detail/widgets/subtask_part_widget.dart
@@ -5,7 +5,7 @@ import "package:theme_prego/module_prego.dart";
 import "../../../core/extensions/build_context_x.dart";
 import "../../../core/routing/app_router.dart";
 import "../../../core/routing/current_project_name.dart";
-import "../../../core/widgets/throttled_activity_indicator.dart";
+import "../../../core/widgets/isolated_activity_indicator.dart";
 
 class SubtaskPartWidget extends StatelessWidget {
   final String? projectId;
@@ -102,7 +102,7 @@ class SubtaskPartWidget extends StatelessWidget {
     SessionStatusBusy() || SessionStatusRetry() => SizedBox(
       width: 16,
       height: 16,
-      child: ThrottledActivityIndicator(
+      child: IsolatedActivityIndicator(
         strokeWidth: 2,
         color: prego.colors.bgBrandSolid,
       ),
